% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/liber-package.R
\docType{package}
\name{liber−package}
\alias{liber−package}
\alias{liber−package-package}
\title{liber}
\description{
Package contains solution of three tasks from homework 6 of
Advanced Data Analysis Software Developmnet with R course
}

