% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simplify2array1}
\alias{simplify2array1}
\title{simplify2array1}
\usage{
simplify2array1(x)
}
\arguments{
\item{list}{of numeric vector}
}
\value{
a numeric matrix
}
\description{
Function simplify2array1 mimic the behaviour of simplify2array R function
if the dimensions of input numeric vectors from list are equal function transforms vectors into
one numeric matrix. Otherwise function throws an error.
}

