% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mode}
\alias{mode}
\title{mode}
\usage{
mode(x)
}
\arguments{
\item{x}{- integer vector}
}
\value{
the most frequently occurring value
}
\description{
function mode determine the most frequently occurring value in an integer
vector
}

